backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
python-dotenv==1.0.1
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.1
Pillow==11.1.0
stripe==12.1.0
